backend/api/pyproject.toml
[project]
name = "aether-api"
version = "0.1.0"
description = "AETHER AI FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.4",
  "celery[redis]>=5.5.3",
  "email-validator>=2.2.0",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.10.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "redis>=6.4.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "asyncpg>=0.30.0",
  "sse-starlette>=2.3.6",
  "uvicorn[standard]>=0.35.0",
  "websockets>=15.0.1"
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.4.1",
  "ruff>=0.12.10",
  "mypy>=1.17.1"
]
